Renamed files to upper camel case convention
