+++ test-node/test.js
@@ -1,4 +1,4 @@
-const DeepstreamClient = require( '../src/client' )
+const DeepstreamClient = require( '../src/Client' )
 const ds = DeepstreamClient( 'ws://localhost:6020' )
 ds.on( 'connectionStateChanged', ( a,b) => {
 	console.log( a,b)
