+++ test-page-webpack/index.js
@@ -1,4 +1,4 @@
-window.deepstream = require('../src/client');
+window.deepstream = require('../src/Client');
 window.ko = require('../test-page/lib/ko.js');
 
 require('../test-page/js/event-view-model.js');
